more adjustments
