+++ web/src/app/(main)/layout.tsx
@@ -1,33 +1,11 @@
 'use client';
-import {
-    faDiscord,
-    faGithub,
-    faPatreon,
-    faTwitter,
-} from '@fortawesome/free-brands-svg-icons';
-import { IconDefinition } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import NextLink from 'next/link';
+import { Box } from '@mui/material';
 import { ReactNode } from 'react';
 import CookieConsent from 'react-cookie-consent';
 import { ToastContainer } from 'react-toastify';
-import Header from '../../components/header/Header';
 import 'react-toastify/dist/ReactToastify.css';
-import {
-    Box,
-    Container,
-    IconButton,
-    Link,
-    Paper,
-    Typography,
-} from '@mui/material';
-
-const icons: { icon: IconDefinition; url: string }[] = [
-    { icon: faGithub, url: 'https://github.com/cjs8487/bingogg' },
-    { icon: faPatreon, url: 'https://www.patreon.com/Bingothon' },
-    { icon: faTwitter, url: 'https://twitter.com/bingothon' },
-    { icon: faDiscord, url: 'https://discord.bingothon.com' },
-];
+import Footer from '../../components/footer/Footer';
+import Header from '../../components/header/Header';
 
 export default function CoreLayout({ children }: { children: ReactNode }) {
     return (
@@ -54,56 +32,7 @@ export default function CoreLayout({ children }: { children: ReactNode }) {
             <Box flexGrow={1} height="100%" display="flex">
                 {children}
             </Box>
-            <Paper component="footer">
-                <Box
-                    sx={{
-                        display: 'flex',
-                        alignItems: 'end',
-                        px: 2,
-                        pt: 2,
-                        pb: 0.5,
-                    }}
-                >
-                    <Box flexGrow={1}>
-                        <Box
-                            sx={{
-                                display: 'flex',
-                                alignItems: 'center',
-                                columnGap: 1,
-                                mb: 0,
-                            }}
-                        >
-                            {icons.map(({ icon, url }) => (
-                                <IconButton
-                                    key={icon.iconName}
-                                    href={url}
-                                    LinkComponent={NextLink}
-                                    size="small"
-                                >
-                                    <FontAwesomeIcon
-                                        icon={icon}
-                                        className="fa-fw aspect-square rounded-full px-1 py-1.5 hover:bg-primary-content hover:bg-opacity-40"
-                                    />
-                                </IconButton>
-                            ))}
-                        </Box>
-                        <Typography variant="caption">
-                            © Copyright 2024 - 2024 Bingothon | All Rights
-                            Reserved |{' '}
-                            <Link
-                                href="/legal/privacy"
-                                component={NextLink}
-                                className="underline"
-                            >
-                                Privacy Policy
-                            </Link>
-                        </Typography>
-                    </Box>
-                    <Typography variant="caption">
-                        bingo.gg v{process.env.version}
-                    </Typography>
-                </Box>
-            </Paper>
+            <Footer />
             <ToastContainer />
         </Box>
     );
